requirements.txt
streamlit~=1.42.0
opencv-python~=4.10.0.84
pandas~=2.2.2
plotly~=5.11.0
protobuf~=4.25.6
scipy~=1.15.1
numpy~=1.26.4
pillow~=11.1.0
tf_keras~=2.18.0